Factor out loaders + add inter/intra catalogs harvest parallelization option.
