+++ README.md
@@ -4,6 +4,7 @@ Spring Boot Server
 
 
 ## Overview  
+
 This server was generated by the [OpenAPI Generator](https://openapi-generator.tech) project.
 By using the [OpenAPI-Spec](https://openapis.org), you can easily generate a server stub.
 This is an example of building a OpenAPI-enabled server in Java using the SpringBoot framework.
@@ -16,17 +17,26 @@ You can view the api documentation in swagger-ui by pointing to
 https://localhost:8443/
 
 ## Configuration setup hints
+
 This back-end is only functional if an application.properties file is added to ```/src/main/resources```.
 This file should contain at least the following:
+
 ```
 springfox.documentation.swagger.v2.path=/api-docs
+
 spring.jackson.date-format=org.openeo.spring.RFC3339DateFormat
 spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS=false
 spring.h2.console.enabled=true
+spring.h2.console.settings.web-allow-others=true
+spring.autoconfigure.exclude = org.keycloak.adapters.springboot.KeycloakAutoConfiguration
+spring.datasource.jdbc=jdbc:h2:/path/to/openeo.db;DB_CLOSE_DELAY=-1
 spring.datasource.username=my_username
 spring.datasource.initialization-mode
+spring.security.filter.order=5
 
+server.tomcat.port=8081
 server.port=8443
+
 server.ssl.key-store-type=PKCS12
 server.ssl.key-store=classpath:my_keystore.p12
 server.ssl.key-store-password=my_keystore_password
@@ -40,11 +50,13 @@ org.openeo.tmp.dir=tmp/
 org.openeo.tmp.file.expiry=60
 org.openeo.file.expiry=1
 org.openeo.querycollectionsonstartup=true
+org.openeo.parallelizedHarvest=true
 
 org.openeo.oidc.configuration.endpoint=https://my_keycloak.url/auth/realms/openeo/
 org.openeo.oidc.providers.list=classpath:oidc_providers.json
 
 org.openeo.wcps.endpoint=http://my_wcps_server:8080
+org.openeo.wcps.endpoint.version=2.0.1
 org.openeo.wcps.provider.name=My Company
 org.openeo.wcps.provider.url=http://www.my-company.url
 org.openeo.wcps.provider.type=host
@@ -65,7 +77,9 @@ org.openeo.udf.r.endpoint=http://my_openeo_R_udf_service.url
 org.openeo.udf.dir=/my/udf/working/directory/
 org.openeo.udf.importscript=/my/udf/import/script/import_udf.sh
 ```
-Further files needed are for connection with keycloak: keycloak.json
+
+Further files needed are for connection with keycloak: `keycloak.json`
+
 ```
 {
 	"realm": "my_realm",
@@ -88,7 +102,9 @@ Further files needed are for connection with keycloak: keycloak.json
 	}
 }
 ```
-and for the support of default client id configuration: oidc_providers.json
+
+and for the support of default client id configuration: `oidc_providers.json`
+
 ```
 {
 	"providers": [
@@ -126,7 +142,8 @@ and for the support of default client id configuration: oidc_providers.json
 ```
 
 ## Logging
-All logging can be controlled through log4j2.
+
+All logging can be controlled through **log4j2**.
 For tweaking of log level and file output modify
-log4j2.xml in ```/src/main/resources```
+`log4j2.xml` in *./src/main/resources*.
