+++ src/main/java/org/openeo/spring/api/loaders/ICollectionParser.java
@@ -1,0 +1,11 @@
+package org.openeo.spring.api.loaders;
+
+import java.util.concurrent.Callable;
+
+import org.openeo.spring.model.Collection;
+
+/**
+ * Tagging interface for callable tasks that return a collection as result.
+ */
+public interface ICollectionParser extends Callable<Collection> {}
+
