+++ src/main/java/org/openeo/spring/api/loaders/ICollectionsLoader.java
@@ -1,0 +1,16 @@
+package org.openeo.spring.api.loaders;
+
+import java.util.concurrent.Callable;
+
+import org.openeo.spring.model.Collections;
+import org.openeo.spring.model.EngineTypes;
+
+/**
+ * Base interface for all loader tasks.
+ */
+public interface ICollectionsLoader extends Callable<Collections> {
+
+    /** Returns the type of engine from where the collections are fetched. */
+    EngineTypes getEngineType();
+
+}
