+++ src/main/java/org/openeo/spring/api/loaders/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains callable tasks that
+ * can be used to load collections/cubes from
+ * different sources, being it a remote service
+ * or a local catalogue file.
+ */
+package org.openeo.spring.api.loaders;
